Added Raised logic
Made it so tiles are either raised or lowered
Made it so tiles render differently depending on if they are raised or
lowered and if they are the first on the column (e.g. if a tile is
raised but the one 'below' it is raised, then only the lower one will
render raised).